pyproject.toml
[project]
name = "microservice"
version = "1.0.0"
description = "Microservice Template"
authors = [
    {name = "",email = "example@example.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "fastapi-pagination (>=0.14.2,<0.15.0)",
    "fastapi-filter (>=2.0.1,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)"
]

[tool.poetry]
packages = [{include="app"}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.13.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
